Ensuring proper urls in sitemap + Updating some packages
baseUrl can't stay "/" because the generated sitemap needs proper urls and can't be changed to the hosting domain because of Netlify's preview and branch builds. This ensures that hugo receives the right baseUrl as parameter: URL for production, DEPLOY_PRIME_URL for branches & preview. Should've separated it out but I also updated some packages