Fix for loading non-svg images in CMS
In the CMS, I was not able to get non-svg images to load using getAsset for the longest time, but simply entry.getIn seems to do the trick.  The Jumbotron loads now, and I did not need to use the relative file path hack that was in the original file.  I think it has to do with getAsset returning a promise, but I am not 100% sure.  I hope that this fixes any image rendering issues that have popped up.